[Project] Move the job codes into Job directory
This commit moves the job-related code to the Job dir.

ONE-vscode-DCO-1.0-Signed-off-by: Jiyoung Yun <jy910.yun@samsung.com>
