+++ src/Project/JobQuantize.ts
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-import {Job} from './Job';
+import {Job} from '../Job/Job';
+import {ToolArgs} from '../Job/ToolArgs';
+
 import {JobBase} from './JobBase';
-import {ToolArgs} from './ToolArgs';
 
 export class JobQuantize extends JobBase {
   // TODO add quantize options
