+++ src/Tests/Project/JobImportBCQ.test.ts
@@ -15,7 +15,7 @@
  */
 
 import {assert} from 'chai';
-import {Job} from '../../Project/Job';
+import {Job} from '../../Job/Job';
 import {JobImportBCQ} from '../../Project/JobImportBCQ';
 
 suite('Project', function() {
