+++ src/Tests/Project/JobImportTFLite.test.ts
@@ -15,7 +15,7 @@
  */
 
 import {assert} from 'chai';
-import {Job} from '../../Project/Job';
+import {Job} from '../../Job/Job';
 import {JobImportTFLite} from '../../Project/JobImportTFLite';
 
 suite('Project', function() {
