+++ src/Tests/Project/JobPack.test.ts
@@ -15,7 +15,7 @@
  */
 
 import {assert} from 'chai';
-import {Job} from '../../Project/Job';
+import {Job} from '../../Job/Job';
 import {JobPack} from '../../Project/JobPack';
 
 suite('Project', function() {
