+++ src/Tests/Toolchain/JobConfig.test.ts
@@ -17,7 +17,7 @@
 import {assert} from 'chai';
 
 import {Command} from '../../Backend/Command';
-import {Job} from '../../Project/Job';
+import {Job} from '../../Job/Job';
 import {JobConfig} from '../../Toolchain/JobConfig';
 
 suite('Project', function() {
